+++ tardis/transport/montecarlo/modes/nonhomologous/rad_packet_transport.py
@@ -19,10 +19,10 @@ from tardis.transport.montecarlo.estimators.estimators_bulk import (
 from tardis.transport.montecarlo.estimators.estimators_line import (
     EstimatorsLine,
 )
-from tardis.transport.montecarlo.packets.radiative_movement import (
+from tardis.transport.montecarlo.packets.movement import (
     increment_packet_cell_index as increment_packet_cell_index,
 )
-from tardis.transport.montecarlo.packets.radiative_movement import (
+from tardis.transport.montecarlo.packets.movement import (
     move_r_packet as shared_move_r_packet,
 )
 from tardis.transport.montecarlo.packets.radiative_packet import (
