+++ tardis/transport/montecarlo/tests/test_montecarlo.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 
 import tardis.transport.montecarlo.modes.classic.rad_packet_transport as r_packet_transport
-import tardis.transport.montecarlo.packets.radiative_movement
+import tardis.transport.montecarlo.packets.movement
 import tardis.transport.montecarlo.packets.radiative_packet as radiative_packet
 import tardis.transport.montecarlo.utils as utils
 from tardis import constants as const
@@ -179,7 +179,7 @@ def test_move_packet_across_shell_boundary_emitted(
     energy = 0.9
     packet = radiative_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    tardis.transport.montecarlo.packets.radiative_movement.increment_packet_cell_index(
+    tardis.transport.montecarlo.packets.movement.increment_packet_cell_index(
         packet, delta_shell, no_of_shells
     )
     assert packet.status == radiative_packet.PacketStatus.EMITTED
@@ -198,7 +198,7 @@ def test_move_packet_across_shell_boundary_reabsorbed(
     energy = 0.9
     packet = radiative_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    tardis.transport.montecarlo.packets.radiative_movement.increment_packet_cell_index(
+    tardis.transport.montecarlo.packets.movement.increment_packet_cell_index(
         packet, delta_shell, no_of_shells
     )
     assert packet.status == radiative_packet.PacketStatus.REABSORBED
@@ -217,7 +217,7 @@ def test_move_packet_across_shell_boundary_increment(
     energy = 0.9
     packet = radiative_packet.RPacket(r, mu, nu, energy)
     packet.current_shell_id = current_shell_id
-    tardis.transport.montecarlo.packets.radiative_movement.increment_packet_cell_index(
+    tardis.transport.montecarlo.packets.movement.increment_packet_cell_index(
         packet, delta_shell, no_of_shells
     )
     assert packet.current_shell_id == current_shell_id + delta_shell
